Refactor extensions, generalize UseValidator
