+++ src/Builders/Extensions/UseValidationStrategyExtensions.cs
@@ -1,5 +1,8 @@
 using System;
+using FluentValidation;
 using FluentValidation.Internal;
+using FluentValidation.Results;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace AppAny.HotChocolate.FluentValidation
 {
@@ -42,9 +45,47 @@ namespace AppAny.HotChocolate.FluentValidation
 			this ArgumentValidationBuilder builder,
 			Action<InputValidatorContext, ValidationStrategy<TInput>> validationStrategy)
 		{
-			return builder.UseInputValidators(context => ValidationDefaults.InputValidators.WithStrategy<TInput>(
-				context,
-				strategy => validationStrategy(context, strategy)));
+			return builder.UseInputValidators(async inputValidatorContext =>
+			{
+				var argumentValue = inputValidatorContext
+					.MiddlewareContext
+					.ArgumentValue<TInput>(inputValidatorContext.Argument.Name);
+
+				if (argumentValue is null)
+				{
+					return null;
+				}
+
+				var validatorType = inputValidatorContext.Argument.GetGenericValidatorType();
+
+				var validators = (IValidator[])inputValidatorContext.MiddlewareContext.Services.GetServices(validatorType);
+
+				var validationContext = ValidationContext<TInput>.CreateWithOptions(
+					argumentValue,
+					strategy => validationStrategy(inputValidatorContext, strategy));
+
+				ValidationResult? validationResult = null;
+
+				for (var validatorIndex = 0; validatorIndex < validators.Length; validatorIndex++)
+				{
+					var validator = validators[validatorIndex];
+
+					var validatorResult = await validator
+						.ValidateAsync(validationContext, inputValidatorContext.MiddlewareContext.RequestAborted)
+						.ConfigureAwait(false);
+
+					if (validationResult is null)
+					{
+						validationResult = validatorResult;
+					}
+					else
+					{
+						validationResult.MergeFailures(validatorResult);
+					}
+				}
+
+				return validationResult;
+			});
 		}
 	}
 }
